fixaps per code review:
* README with more clear explenations and examples
* Avoid copying local file into temp file
* move insert_hash_in_path as method for DumperBase
* insert hash in file path in more relaible way
* avoid code duplication
