+++ datapackage_pipelines/utilities/resources.py
@@ -1,6 +1,3 @@
-import os
-
-
 def is_a_url(path):
     return (path is not None and isinstance(path, str) and
             (path.startswith('http://') or
@@ -32,22 +29,6 @@ def get_path(descriptor):
             return None
     assert path is None, '%r' % path
     return None
-
-
-def insert_hash_in_path(descriptor, hash):
-    path = descriptor.get('path')
-    if isinstance(path, str):
-        path_list = path.split('/')
-        hashed = os.path.join(hash, path_list.pop(-1))
-        path_list.append(hashed)
-        descriptor['path'] = '/'.join(path_list)
-    if isinstance(path, list):
-        if len(path) > 0:
-            path = path[0]
-            path_list = path.split('/')
-            hashed = os.path.join(hash, path_list.pop(-1))
-            path_list.append(hashed)
-            descriptor['path'] = ['/'.join(path_list)]
 
 
 PATH_PLACEHOLDER = '.'
